+++ Unity/Assets/Model/Module/Message/MessageInfo.cs
@@ -2,10 +2,10 @@
 {
 	public struct MessageInfo
 	{
-		public ushort Opcode { get; }
+		public uint Opcode { get; }
 		public object Message { get; }
 
-		public MessageInfo(ushort opcode, object message)
+		public MessageInfo(uint opcode, object message)
 		{
 			this.Opcode = opcode;
 			this.Message = message;
