+++ Unity/Assets/Model/Module/Message/OuterMessageDispatcher.cs
@@ -2,7 +2,7 @@
 {
 	public class OuterMessageDispatcher: IMessageDispatcher
 	{
-		public void Dispatch(Session session, ushort opcode, object message)
+		public void Dispatch(Session session, uint opcode, object message)
 		{
 			// 普通消息或者是Rpc请求消息
 			MessageInfo messageInfo = new MessageInfo(opcode, message);
