+++ Unity/Assets/Model/Module/Message/SessionCallbackComponent.cs
@@ -5,7 +5,7 @@ namespace ET
 {
 	public class SessionCallbackComponent: Entity
 	{
-		public Action<Session, ushort, MemoryStream> MessageCallback;
+		public Action<Session, uint, MemoryStream> MessageCallback;
 		public Action<Session> DisposeCallback;
 
 		public override void Dispose()
